requirements.txt
gradio==5.49.1
langchain==1.0.2
langchain-chroma==0.2.6
langchain-core==1.0.0
langchain-openai==0.3.34
langgraph==1.0.1
langgraph-checkpoint==3.0.0
chromadb==1.1.0
openai==2.1.0
python-dotenv==1.1.1